Add authorization functionality and auth checks for all todo list modifying methods

+++ todolist-api.js
@@ -1,28 +1,51 @@
 const express = require('express');
-const crypto = require("crypto");
+const sessions = require('express-session');
 const Joi = require('joi')
 const { init, addTask, getTasks, getTaskByID, updateTask, deleteTask } = require('./db.js');
 
 const app = express();
 const port = process.env.port || 3000;
 
+const oneDay = 1000 * 60 * 60 * 24;
+app.use(sessions({
+    secret: "secretsarecool5783457489758923765",
+    saveUninitialized:true,
+    cookie: { maxAge: oneDay },
+    resave: false
+}));
+
 const taskSchema = Joi.object().keys({
     name: Joi.string().length(32),
     description: Joi.string()
 })
 
+// Simulated unencrypted username/password DB
+let users = [{username: Jerrr, password: myPassword}, {username: brokeJT, password: ILikeToast2000}];
+
 app.get('/', (req, res) => {
     res.send("Todolist API");
 })
 
+app.post('/login', (req, res) => {
+    let userInfo = users.find(user => user.username === req.body.username);
+    if (!userInfo || userInfo.password !== req.body.password) {
+        res.sendStatus(401);
+    } else {
+        let session=req.session;
+        session.userid=req.body.username;
+        console.log(req.session)
+    }
+});
+
 app.get('/tasks', (req, res) => {
-    // Get tasks from DB
-    getTasks()
+    // Get tasks from DB, whose descriptions contain filterString if it's included as a parameter
+    getTasks(req.query.filterString)
         .then((tasks) => {
             tasks = tasks.map((task) => ({
                 id: task._id,
                 name: task.name,
-                description: task.description
+                description: task.description,
+                user: task.user
             }));
 
             res.json(tasks);
@@ -39,7 +62,8 @@ app.get('/tasks/:id', (req, res) => {
             let mappedTask = {
                 id: task._id,
                 name: task.name,
-                description: task.description
+                description: task.description,
+                user: task.user
             }
 
             res.json(mappedTask);
@@ -59,7 +83,11 @@ app.post('/tasks', (req, res) => {
         res.sendStatus(400);
         return
     }
-
+    if (!req.session.userid) {
+        res.status(401).send("You need to be logged in!");
+        return
+    }
+    task.user = req.session.userid;
     addTask(task)
         .then(() => res.sendStatus(200))
         .catch((err) => {
@@ -69,17 +97,47 @@ app.post('/tasks', (req, res) => {
 });
 
 app.put('/tasks/:id', (req, res) => {
-    updateTask(req.params.id, req.body)
-        .then(() => res.sendStatus(200))
-        .catch((err) => {
-            console.log(err);
-            res.sendStatus(500);
+    const task = req.body;
+
+    const result = taskSchema.validate(task);
+    if(result.error) {
+        console.log(result.error);
+        res.sendStatus(400);
+        return
+    }
+    getTaskByID(req.params.id)
+        .then((task) => {
+            if (task.user === req.session.userid) {
+                updateTask(req.params.id, req.body)
+                    .then(() => res.sendStatus(200))
+                    .catch((err) => {
+                        console.log(err);
+                        res.sendStatus(500);
+                    })
+            } else {
+                res.status(401).send("You need to be logged in!");
+            }
         })
+        .catch((err) => {
+        console.log(err);
+        res.sendStatus(500);
+    })
 });
 
 app.delete('/tasks/:id', (req, res) => {
-    deleteTask(req.params.id)
-        .then(() => res.sendStatus(200))
+    getTaskByID(req.params.id)
+        .then((task) => {
+            if (task.user === req.session.userid) {
+                deleteTask(req.params.id)
+                    .then(() => res.sendStatus(200))
+                    .catch((err) => {
+                        console.log(err);
+                        res.sendStatus(500);
+                    })
+            } else {
+                res.status(401).send("You need to be logged in!");
+            }
+        })
         .catch((err) => {
             console.log(err);
             res.sendStatus(500);
